Create a new blogdown post (with tweaks)

This is a modified version of blogdown::new_post for extra taxonomies.

new_post(
  title,
  kind = "",
  open = interactive(),
  author = getOption("blogdown.author"),
  categories = NULL,
  tags = NULL,
  packages = NULL,
  shows = NULL,
  date = Sys.Date(),
  file = NULL,
  slug = NULL,
  title_case = getOption("blogdown.title_case"),
  subdir = getOption("blogdown.subdir", "post"),
  ext = getOption("blogdown.ext", ".md")
)

Arguments

title
The title of the post.

kind
The archetype to use.

open
Whether to open the post in the editor, defaults to TRUE if in an interactive setting.

author
The author of the post.

categories
A character vector of category names.

tags
A character vector of tag names.

packages, shows
A character vector of package names or show names.

date
The date of the post.

file
The filename of the post. By default, the filename will be automatically generated from the title by replacing non-alphanumeric characters with dashes, e.g. title = 'Hello World' may create a file content/post/2016-12-28-hello-world.md. The date of the form YYYY-mm-dd will be prepended if the filename does not start with a date.

slug
The slug of the post. By default (NULL), the slug is generated from the filename by removing the date and filename extension, e.g., if file = 'post/2015-07-23-hi-there.md', slug will be hi-there. Set slug = '' if you do not want it.

title_case
A function to convert the title to title case. If TRUE, the function is tools::toTitleCase()). This argument is not limited to title case conversion. You can provide an arbitrary R function to convert the title.

subdir
If specified (not NULL), the post will be generated under a subdirectory under content/. It can be a nested subdirectory like post/joe/.

ext
The filename extension (e.g., .md, .Rmd, or .Rmarkdown). Ignored if file has been specified.